abstract factory classes
